add routing functionality to Grid
It relies on an optional property *routing_info* of Grid to be set using a numpy array featuring the direction of the movement departing each grid cell in the domain.

If this info is given, then the component which is given this Grid can use the *route* method to move the values in the array given do their immediate neighbour according to the direction set in the *routing_info* property.