enforce lower_left to upper_right spacedomain in Grid
Since it is important for the routing functionality of a Grid, more checks were added to make sure that a Grid is regular (constant spacing between coordinates), has coordinates in their bounds, is directed positively (i.e. northwards and eastwards, i.e. from lower left to upper right), has coordinates within axis limits.

It supports a cyclic coordinate for longitude, i.e. wrap around is allowed (e.g. [179, 180, -179] is a valid positive sequence of longitudes). It does not support a cyclic coordinate for latitude (which would be a edge case anyway, as a global run would not include a pole anywhere else than on the edge of the domain).