make use of enforced lower left to upper right in Grid.route
In the previous implementation, the routing functionality was already making an assumption on the orientation of the domain. But this assumption was that higher latitude would be at the top of the array.

With the latest enforcement of lower left to upper right domain for a Grid (commit 4a15ab1011a71ddd5c8a9e9a820e5ea2ded3650b), a dimension coordinate sequence will start with lower latitudes (possibly negative) towards higher latitudes (i.e. in a northwards direction), so in a numpy array, this means that the first row will start with the lower latitudes (southernmost) at the top of the array.

So this commit simply adjusts the algorithm where appropriate to take into account this flipping for the latitude axis.