add decimal rounding before space subset
`cf.wi` (like numpy.__le__ and numpy.__ge__) does not allow for a tolerance on the comparison operators, which means that sometimes a coordinate in the subset is dropped due to floating point error problems.

e.g., the dummy test with a resolution of 0.2 deg raised an error because the last coordinate evaluates as 54.900000000000055, with a cf.wi(51,1, 54.9) the last coordinate is dropped in the space subset, which yields an error in the Grid.is_space_equal_to(field).