rename variables stored in file 'record' instead of 'output'
The term output was ambiguous, because it is the word that was used to refer to those variables produced by a component, that are not already part of the outwards (variables destinated to other components), and to those variables that are requested to be (optionally aggregated and) stored in a NC file.

So renamed the latter, i.e. those variables requested to be stored in a file, as 'records' rather than 'outputs'. Also renamed the 'output_directory' to a 'saving_directory' for the same reason.