create a quicker set of tests 'basic'
This is done by distinguishing the series of tests in test_model into 'basic' and 'advanced'. Advanced inherits from basic. Advances is used by 'run_all_tests.py' which means it is in effect unchanged in scope with before. While basic is used for same time/same space and for diff time/diff space in 'run_basic_tests.py' alongside all doctests in order to provide a quicker set of tests to run, while remaining rather good in terms of coverage (even if not quantified so far).