fix bug when only long_name in config file
The select item in 'to_config' method is populated with the long_name attribute rather than "long_name=...", which is not finding any field with cf.read(field, select="..."), as it needs the leading "long_name=" if it is not a standard_name.

So now the 'to_config' method uses the field.identity() method, which will either return the standard_name, or the long_name but preceded with the leading "long_name=".

This needed modifying the '_get_dict_variables_from_file' to support the leading "long_name=" in the name_mapping. For consistency, it also supports the leading "standard_name=", even if not strictly necessary for the select of cf.read.

Also added some examples in docstrings for doctest to run.