change type to field for setting land_sea_mask and flow_direction
By expecting a `cf.Field` rather than a `numpy.ndarray`, the data can be checked on its space dimensions, not only the shape of the array, this way it is guaranteed that the data is for where it is meant to be.

Also rename the 'routing_info' to 'flow_direction' attribute, which is more explicit and compliant with existing datasets.