add tests using land_sea_mask and flow_direction
Dummy SurfaceLayer is the component chosen as the one requiring both land_sea_mask and flow_direction. The dummy data itself does not change the outcome of the runs because land_sea_mask is set to 1 everywhere (i.e. land everywhere) and flow_direction is set to 0 everywhere (i.e. no flow/stagnant). This has the merit to test that the required special SpaceDomain attributes are provided, and to use the routing functionality of Grid.

In addition, it also tests the subspacing capacity on these datasets, because these ancillaries are created as global datasets.