add BritishNationalGrid
BritishNationalGrid (BNG) is the first cartesian coordinate reference system SpaceDomain to be included. But since its eastings/northings are projected into 'true' (WGS84) lat/lon coordinates, the same spherical regridding functionality will work (although the cartesian regridding would work too, it is preferable to avoid special cases).

The coordinate reference for the BNG is not customisable, in order to remove the burden on the user, and it uses all the parameters found in EPSG:27700.