add routing around limits for axes with contiguous limits
For LatLonGrid and RotatedLatLonGrid, their limits being contiguous, their `route` methods should not consider the edge of their arrays as the edge of their domains if these have global coverage. This commit uses new class attibutes on contiguity for Z/Y/X to determine whether flow direction towards outside the array should be rolled around or not.

Also rename class attributes on cyclicity as wrap_around to avoid confusion with these new contiguity attributes, and to explicit that these are used to determine whether wrap around on domain definitions are allowed or not. For global Grids, it is only allowed for the X-axis (not the Y-axis which is avoided due to complications around the poles).