remove unsupported domain when all diffs are negative
If the domain contains only one cell, and this cell happens to be wrapping around the limits of a cyclic dimension, the catch intended to catch earlier invalid domains erroneously catches this negative value as representing an invalid domain, which is not the case. So this commit removes this early catch, and let the rest of the algorithm do the job (i.e. too many negative values will raise an error later on).