allow for construction of Grids with same start/end extent
For axes where wrap around is permitted, the constructor method `from_extent_and_resolution` was raising an error when start and end of extent were equal. It is preferable to assume that the user intends on constructing a wrap-around domain on this axis, so this is what this commit does. If they are equal, end is added a full rotation along this axis, just like it was the case when end was smaller than start.