break down files clean up in model test and add missing file catches
Breaking up the clean up between dump/record/config files will allow for more flexibility in the tests in the future.

Also the change in filename in 575db2b6a9bd824535b58528ac2df34d75442c80 was not followed by a change in the clean up search, it was still looking for files with '_out' in them and not '_records.

Finally, it was using the saving_directory of exchanger (inherited from model) to look for those files (which was fine in these tests because the same directory is used for everything), but the saving_directory of the component files can poentielly be different from the model one, so needed to broaden the search.