remove clean up of 2nd simulator
When only with data/nullcomponents, the second simulator in this test will have no exchanger because the resume method will abort due to no "real" component in the model.

For cases with "real" components, given that simulator_1 and simulator_2 have the same identifier, the clean up with the former is sufficient to remove the files.