Integrate Artemis and RFM as science components (#22)
* add Artemis as surfacelayer and subsurface components

* add RFM as openwater component

* update docs to include ARTEMIS+RFM

also largely restructure the docs to include guidance for users and for contributors

* fix problem with duplicated image in docs

using the same image file resulted in duplicating the file during the docs build, so remove the use of the main diagram image in the preparation section, and directly refer to the one on the overview page

* add setup requirements in text files

* create pyproject.toml to explicit build requirements

i.e. need for setuptools and wheel

* widen scope of context manager to catch np overflow warnings in ArtemisSL

These warnings are due to the _FillValue of the masked numpy arrays. As indicated here https://stackoverflow.com/a/12515776/7757430 it seems that the calculations are applied to the entire array before reapplying the mask.