update mock example for component preparation to use 'divisions'
This namely shows that 'divisions' is appended as last axis for the component state array.

While at it, this also fixes the mistake with climatologic which was not index, so the mock example would raise a broadcasting error. This is now indexed properly using the first axis of the input array (this is where the time dimension for climatologic input is).

There were several errors in units. These are fixed now, and it should pass a dimensional analysis. And the names of the component definition are shortened to clarify the code.