add support for non-scalar component parameter values
The supported types for non-scalar values are:
- sequence (array, units)
- cf.Data
- cf.Field

 If they are cf.Field, there is a subset performed on the values using the component SpaceDomain, and then a comparison to see if spatial metadata match.
If they are cf.Data or sequence (array, units), there can be no spatial metadata match or subset because they are not spatially explicit, so only the shape of the arrays are compared with the shape of the component SpaceDomain.

For scalar component values, a corresponding array of the size of the component SpaceDomain is created and filled with the scalar. This way, a component will always know what will be the shape of the parameter it is receiving, which is particularly important for Fortran and C components.

The tests are updated to include parameters provided as a cf.Field (in addition to existing cf.Data and sequence [scalar, units]).