make SpaceDomain an abstract base class
This is to enforce the compulsory attributes and methods that need to be implemented if a new kind of SpaceDomain is developed.

Also removed duplicates in some of the construction methods of the different Grids, they could all inherit most of the functionality from superclass Grid.

Also extended the tests to run them on RotatedLatLonGrid and BritishNationalGrid too.