add time slice on input data reading
By creating Variable objects stored in the DataSet in place of the cf.Field, the data load on disk can be controlled to retrieve more than one time step at a time, in order to reduce the number of times files are opened/closed.

Variable is declined into a DynamicVariable (where there is time slicing), and into a StaticVariable and a ClimatologicVariable (where there is no time slicing).