add a test to the "basic tests suite" to include Data/Null component
While the full set of real/data/null components is tested in the "advanced test suite", not having a test using DataComponent and NullComponent was not a good idea, because they behave slightly differently than a real Component in some regard.

So a test inspired by the one in "advanced" is added in "basic" and works with the combination (SL: data, SS: real, OW: null).