use custom io_slice parameter value for tests
Chose a value of 10 because it is both lower than simulation period for main run (12 days), and greater than spinup period (6days), effectively resulting in 6 day record stream beats in both cases but for different algorithmic reasons (i.e. shortened to keep a multiple of simulation period in the former case, shortened/limited to/by the spin up period in the latter case).