review properties in component definition
Old *info* properties used to return references to the internal dictionaries used to define the components, which was dangerous because the user could modify them inadvertently, so they are replaced by *metadata* properties that return a string displaying the content of these dictionaries.

In addition, *info* properties are kept for inwards/outwards (because required for Exchanger and Data/NullComponent), but they return a deep copy of the dictionaries to avoid accidentally modifying them internally.