reorganise parameters between coordinate conversion and datum
The split of coordinate reference into coordinate conversion and datum is an artefact of the CF data model, so when `cf-python` reads them in, there was a reshuffle of some of the properties between coordinate conversion and datum for that very reason. So now `cm4twc` follows the same "rules" as `cf-python` regarding which parameters belongs to which object.

Also, due to this artefact, if both coordinate conversion and datum feature a 'unit' parameter, since 'unit' belongs to coordinate conversion, one of the 'unit' will be lost when reading a field with `cf-python`. However, 'unit' is not actually required by the CF conventions as false_easting|northing takes the same unit as the projection_x|y_coordinate, and longitude_of_prime_meridian is mandated to be in degrees, so we can simply remove them from `cm4twc`'s SpaceDomains, and if it is in the data field, the parameter is deleted before the equality check.

Also, this adds some missing 'grid_mapping_name' that should have been in 37723504fea010262172e4845a62d6029c6dda0e