use explicit attributes rather than construct to get CF objects
resolve #51

This was namely problematic for `BritishNationalGrid` which has both projection_y|x_coordinate (i.e. dimension coordinates) and latitude|longitude (i.e. auxiliary coordinates), so that querying through construct does not work as it does not know which one of the dimension coordinate or auxiliary coordinate it should return.