add cell_area property to spacedomain
resolve #61

A cell_area property similar in design to land_sea_mask and flow_direction is added to `SpaceDomain`. The cell_area can be provided by the user as a `cf.Field` and for `Grid`, the cell_area is automatically computed if the property "getter" is called and no values have already been given/computed.