add component requirement property for cell area
Since the implementation of cell_area property of spacedomain in 5b038bce33d9de68fa7f1e182e88bc9c936f687b, a component could want to use this property, but there was no way for the component contributor to tell the framework this data was required. This is now the case following the same approach as for land_sea_mask and flow_direction.