fix display of model components
Since #46, the component contributions are expected to inherit not only the framework component classes, but their names as well (i.e. SurfaceLayerComponent, SubSurfaceComponent, OpenWaterComponent). When "printing" a Model instance, the information displayed became irrelevant, as the class name would not allow to tell which science component is actually being used. This is now fixed by displaying the full module name containing the component class.