review spatial resolution constraint in Compass
The new constraint is that: "each component temporal resolution must be an integer multiple of the other components' resolutions".

So a space domain compatibility check in `Compass` is revised accordingly. The `Grid` is fitted with a new method `is_matched_in` to determine whether one grid is "contained" in another.

Also tests are added in the docstring to make sure that the check functions and remains functioning as expected.

Also new Grid method `is_matched_in` is added to the API reference.