update dummy component resolutions in tests to comply with new constraints
The component temporal resolutions used to be {SL: daily, SS: 3daily, OW: 2daily}, but 3 is not a multiple of 2. They are now {SL: daily, SS: 4daily, OW:2daily}. The simulation period is also extended to 16 days, so that each spin up period is half of the simulation period.

The component spatial resolutions used to be {SL: 1deg, SS: 0.5deg, OW: 0.2deg}, but 0.5 is not a multiple of 0.2. They are now {SL: 1deg, SS: 0.5deg, OW:0.25deg}.

The expected records needed to be modified to take into account these modifications. These were manually calculated, and then compared against the results using the framework to make sure these new values are robust - they are essential because they form the basis for checking that the framework has not been broken.

The input files and the substitute data files needed to also be modified to comply with the resolutions and temporal extent changes.

The YAML configuration files are also modified to reflect the changes in the tests themselves.