drop vertical dimension for land_sea_mask, flow_direction, cell_area
since these 3 attributes of a space domain are physically only meaningful horizontally, even if `SpaceDomain` has a vertical axis.

So the compatibility check between the `SpaceDomain` and the data field is done on the horizontal `Grid` only, the data field is squeezed along Z, and the Z axis is dropped in the space shape of the processed data.