add comment as cut-off tag for docs
Because the wording of the last paragraph may change, I have added a comment that is now used as the cut-off point so that not all of the README file is displayed on the homepage of the docs (e.g. here the link to the docs does not make sense to have on the homepage of the said docs).